requirements.txt
pandas
numpy
yfinance
fredapi
pytrends
requests
beautifulsoup4
selenium
refinitiv-data
xgboost
lightgbm
scikit-learn
matplotlib
seaborn
plotly
scipy
statsmodels
python-dotenv
openpyxl
jupyter
